+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### [13.19.2](https://kaos.sh/ek/13.19.2)
+
+- **`[timeutil]`** Added helper `Since`
+- **`[spellcheck]`** Code refactoring
+- **`[system]`** Code refactoring
+
 ### [13.19.1](https://kaos.sh/ek/13.19.1)
 
 - **`[ansi]`** Code refactoring